removing more such code
