+++ compiler/src/scope/concrete/core.rs
@@ -1,9 +1,6 @@
 use crate::scope::core::AbstractConcreteTypesHandler;
-use crate::scope::core::GenericTypeParams;
 use crate::scope::core::SymbolData;
 use crate::types::core::Type;
-
-use super::callable_registry::CallableConcreteTypesRegistry;
 
 #[derive(Debug, Clone, Copy)]
 pub struct ConcreteTypesRegistryKey(pub usize);
@@ -34,21 +31,6 @@ impl<T: AbstractConcreteTypesHandler> Clone for ConcreteSymbolData<T> {
         ConcreteSymbolData {
             symbol_data: self.symbol_data.clone(),
             index: self.index,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct GenericsSpecAndConcreteTypesRegistry {
-    pub generics_spec: GenericTypeParams,
-    pub concrete_types_registry: CallableConcreteTypesRegistry,
-}
-
-impl GenericsSpecAndConcreteTypesRegistry {
-    fn new(generics_spec: GenericTypeParams) -> Self {
-        GenericsSpecAndConcreteTypesRegistry {
-            generics_spec,
-            concrete_types_registry: CallableConcreteTypesRegistry::default(),
         }
     }
 }
