+++ compiler/src/scope/concrete/mod.rs
@@ -1,3 +1,2 @@
-pub mod callable_registry;
 pub mod core;
-pub mod struct_registry;
+pub mod registry;
